requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.118.0
pandas==2.2.1
matplotlib==3.8.2
seaborn==0.13.0
google-generativeai==0.3.2
textblob==0.17.1
python-dotenv==1.0.1
gspread==5.12.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0 